Refactor: extendedMessages 만드는 로직 외부로 분리
