+++ src/pages/Chats/ChatRoom/index.tsx
@@ -1,34 +1,34 @@
 import { useState, useEffect, useRef } from 'react';
+import { useNavigate, useParams } from 'react-router-dom';
+import { useRecoilState, useRecoilValue } from 'recoil';
 import { MessagesContainer } from './styles';
+import { OODDFrame } from '../../../components/Frame/Frame';
 import TopBar from '../../../components/TopBar';
 import RcvdMessage from './RcvdMessage';
 import SentMessage from './SentMessage';
 import DateBar from './DateBar';
 import ChatBox from './ChatBox';
-import { useRecoilState, useRecoilValue } from 'recoil';
 import BottomSheet from '../../../components/BottomSheet';
-import dayjs from 'dayjs';
-import 'dayjs/locale/ko';
-import ConfirmationModal from '../../../components/ConfirmationModal';
-import { OODDFrame } from '../../../components/Frame/Frame';
-import { MessageDto, ExtendedMessageDto, SentMessageProps, RcvdMessageProps } from '../dto';
-import { ConfirmationModalProps } from '../../../components/ConfirmationModal/dto';
-import Exit from '../../../assets/BottomSheetMenu/Exit.svg';
-import Block from '../../../assets/BottomSheetMenu/Block.svg';
-import { BottomSheetProps } from '../../../components/BottomSheet/dto';
 import BottomSheetMenu from '../../../components/BottomSheetMenu';
+import ConfirmationModal from '../../../components/ConfirmationModal';
+import Modal from '../../../components/Modal';
+import { ExtendedMessageDto } from '../dto';
+import { BottomSheetProps } from '../../../components/BottomSheet/dto';
 import { BottomSheetMenuProps } from '../../../components/BottomSheetMenu/dto';
-import { useNavigate, useParams } from 'react-router-dom';
-import request from '../../../apis/core';
+import { ConfirmationModalProps } from '../../../components/ConfirmationModal/dto';
+import { ModalProps } from '../../../components/Modal/dto';
+import { createExtendedMessages } from './createExtendedMessages';
 import { AllMesagesAtom } from '../../../recoil/AllMessages';
 import { OpponentInfoAtom } from '../../../recoil/OpponentInfo';
-import ProfileImg from '/ProfileImg.svg';
-import Back from '../../../assets/Chats/Back.svg';
-import KebabMenu from '../../../assets/Chats/KebabMenu.svg';
 import { useSocket } from '../../../context/SocketProvider';
 import { ApiDto } from './dto';
-import { ModalProps } from '../../../components/Modal/dto';
-import Modal from '../../../components/Modal';
+import request from '../../../apis/core';
+import Back from '../../../assets/Chats/Back.svg';
+import KebabMenu from '../../../assets/Chats/KebabMenu.svg';
+import Exit from '../../../assets/BottomSheetMenu/Exit.svg';
+import Block from '../../../assets/BottomSheetMenu/Block.svg';
+import dayjs from 'dayjs';
+import 'dayjs/locale/ko';
 
 const ChatRoom: React.FC = () => {
 	const [extendedMessages, setextendedMessages] = useState<ExtendedMessageDto[]>([]);
@@ -60,7 +60,7 @@ const ChatRoom: React.FC = () => {
 		socket.on('AllMessages', (messages) => {
 			setAllMessages(messages.reverse());
 			if (messages.length > messageLengthRef.current) {
-				setIsScroll(true);
+				setIsScroll((prev) => !prev);
 				setIsLoaded(true);
 			}
 		});
@@ -68,7 +68,7 @@ const ChatRoom: React.FC = () => {
 		// 최근 메시지 조회
 		socket.on('latestMessage', (message) => {
 			setAllMessages((prevMessages) => [...prevMessages, message]);
-			setIsScroll(true);
+			setIsScroll((prev) => !prev);
 		});
 
 		// 컴포넌트 언마운트 시 실행
@@ -79,6 +79,12 @@ const ChatRoom: React.FC = () => {
 			}
 		};
 	}, [socket, roomId]);
+
+	// 메시지 렌더링에 필요한 정보 추가
+	useEffect(() => {
+		const temp = createExtendedMessages();
+		setextendedMessages(temp);
+	}, [allMessages]);
 
 	function scrollToBottom(ref: React.RefObject<HTMLDivElement>) {
 		if (ref.current) ref.current.scrollIntoView();
@@ -97,7 +103,7 @@ const ChatRoom: React.FC = () => {
 	useEffect(() => {
 		if (isScroll) {
 			scrollToBottom(chatWindowRef);
-			setIsScroll(false);
+			setIsScroll((prev) => !prev);
 		}
 	}, [isScroll]);
 
@@ -203,60 +209,6 @@ const ChatRoom: React.FC = () => {
 			setIsOpenMenu(false);
 		},
 	};
-
-	// DateBar 표시 여부를 결정하는 함수
-	const isNextDay = (curDate: Date, lastDate: Date): boolean => {
-		const curDateDayjs = dayjs(curDate);
-		const lastDateDayjs = dayjs(lastDate);
-		return !curDateDayjs.isSame(lastDateDayjs, 'day');
-	};
-
-	// 기존 대화 내역에 대한 정보 추가
-	useEffect(() => {
-		const temp: ExtendedMessageDto[] = allMessages.map((message: MessageDto, index: number) => {
-			const prevMessage = index !== 0 ? allMessages[index - 1] : null;
-			const nextMessage = index !== allMessages.length - 1 ? allMessages[index + 1] : null;
-			const formattedTime = dayjs(message.createdAt).format('HH:mm');
-
-			// 채팅의 첫 메시지가 아니고, 날짜가 바뀐 경우 날짜 표시줄 출력
-			let isNewDate = prevMessage !== null && isNextDay(message.createdAt, prevMessage.createdAt);
-
-			// 채팅의 첫 메시지이거나 전송자 또는 날짜가 바뀐 경우 프로필 사진 출력
-			let isFirst = prevMessage === null || prevMessage.fromUser.id !== message.fromUser.id || isNewDate;
-
-			// 가장 마지막 메시지이거나,
-			// 전송자 또는 시간 또는 날짜가 바뀌기 직전인 경우
-			// 메시지 전송 시각 출력
-			let isPrintTime =
-				nextMessage === null ||
-				message.fromUser.id !== nextMessage.fromUser.id ||
-				formattedTime !== dayjs(nextMessage.createdAt).format('HH:mm') ||
-				isNextDay(nextMessage.createdAt, message.createdAt);
-
-			// 채팅의 첫 메시지가 아니고, 전송자가 바뀐 경우 margin-top 추가
-			let isSenderChanged = prevMessage !== null && !isNewDate && prevMessage.fromUser.id !== message.fromUser.id;
-
-			// 보낸 메시지일 경우 sentMessage 속성 추가
-			// 받은 메시지일 경우 rcvdMessage 속성 추가
-			if (message.fromUser.id === userId) {
-				const sentMessage: SentMessageProps = { content: message.content, isSenderChanged, isPrintTime, formattedTime };
-				return { ...message, isNewDate, sentMessage };
-			} else {
-				const rcvdMessage: RcvdMessageProps = {
-					fromUserName: opponentInfo?.name || '알수없음',
-					profilePictureUrl: opponentInfo?.profilePictureUrl || ProfileImg,
-					content: message.content,
-					isFirst,
-					isSenderChanged,
-					isPrintTime,
-					formattedTime,
-				};
-				return { ...message, isNewDate, rcvdMessage };
-			}
-		});
-
-		setextendedMessages(temp);
-	}, [allMessages]);
 
 	return (
 		<OODDFrame>
